Replace LabelWithExternalIcon component with InteractiveLabel
make it more generic and add visual feedback
