+++ src/components/FeesField/GenericContainer.js
@@ -3,7 +3,8 @@
 import React from 'react'
 
 import Box from 'components/base/Box'
-import LabelWithExternalIcon from 'components/base/LabelWithExternalIcon'
+import { InteractiveLabel } from 'components/base/Label'
+import IconExternalLink from 'icons/ExternalLink'
 import { translate } from 'react-i18next'
 import { openURL } from 'helpers/linking'
 import { urls } from 'config/urls'
@@ -11,7 +12,8 @@ import { track } from 'analytics/segment'
 
 export default translate()(({ children, t }: { children: React$Node, t: * }) => (
   <Box flow={1}>
-    <LabelWithExternalIcon
+    <InteractiveLabel
+      Icon={IconExternalLink}
       onClick={() => {
         openURL(urls.feesMoreInfo)
         track('Send Flow Fees Help Requested')
